Changed files to be a hosted server.  Safe revert point.
